CSS Screen now passes button presses to the playerPanels to control their fighter wheels.
In order to clean this up, some changes were made to settingsManager, including a new getControls(playerNum) function, that will spit out an empty controls array if you try to get controls for a player that haven't been set.
Previously, you'd get errors asking for, for example, controls_3 in a 2-player game. Also, refactored some of the ini parsing so now asking for a setting that isn't there will spit out an error message and give a default value (empty string, False, or 0) instead of crashing.

This can still cause fatal errors down the line, but if, for example, you forget to set your "enableWavedashing" flag in your ini file, the game will treat it as if it were there but set to False.