Linked rules menu to game. Tweaked music.
Currently, only stock works, but it does end the game when 1 or fewer
people are left after a player elimination. Music manager now fades out
on selecting a stage, and plays the stage music. Music stops when the
game ends, but exiting early does not reset the music at the moment.