Added hitbox clanking.
If two hitboxes connect, if one is not significantly stronger than the other,
they will cancel. Added an onClank(actor) function to actions, so that
actions can define what happens when they are clanked. As an example,
hitboxie's side special will go into the spindown period of the move if it
clanks.