Skeleton'd AI code.
Player two will now follow player one around. This is to test the AI
input system only, as there is currently no way to disable the AI
player. If you would like to play with two players, please set the human
player to player 3 and do your best to ignore the bot until we rollout
the toggle for him.