Added native controller support (Hell yeah!)
Controllers now work in Project TUSSLE! Rejoice! This one's been an
issue for a long time. It's not done yet, there's a lot of UI and UX
stuff to clear up, but they are functionally completed. You'll need to
set the fighter's controlType to the name of your gamepad, and define
the controls as the following:

Axis inputs are labeled aX = (N,P) where X is the axis number, N is the
input for negative on that axis, and P is the input for positive on that
axis. For example, if axis 0 was the left-right axis, with left being
negative, you would define in the settings.ini: a0 = (left,right)

Expect a Feature Friday on it soon, once more of it is finished.