Action inheritance has been redone. Dynamic Action has been removed.
Okay, so, this one was a doozy and I'm gonna write more than a lazy
commit message for it. Before we had two "top level" action classes,
Action, for pythonic actions, and DynamicAction, for ones generated in
XML. There were a series of "Base Actions" that inherited from Action
and did the basic functionality of the action. Problem was, we'd like
those base actions to work in XML as well. So, basically, we were
rewriting the class hierarchy of the DynamicActions to include the
BaseActions, even if they weren't normal Actions themselves. This turned
out to be a problem, the DynamicActions actually retained ALL classes
you tried to throw at it. It would simultaneously be every class it ever
encountered like some kind of weird code SCP-682.

Removed DynamicAction, moved all of its functionality into Action, and
rewrote all BaseActions to call the new DynamicStuff if they didn't
already. Now DynamicActions properly follow the hierarchy of Action ->
BaseAction -> Generated XML Action.